+++ Pods/Commander/Sources/Commander/ArgumentDescription.swift
@@ -27,6 +27,7 @@ extension ArgumentConvertible {
 }
 
 
+@available(*, deprecated, message: "use array as an ArgumentConvertible")
 public class VariadicArgument<T : ArgumentConvertible> : ArgumentDescriptor {
   public typealias ValueType = [T]
   public typealias Validator = (ValueType) throws -> ValueType
@@ -34,7 +35,6 @@ public class VariadicArgument<T : ArgumentConvertible> : ArgumentDescriptor {
   public let name: String
   public let description: String?
   public let validator: Validator?
-
   public var type: ArgumentType { return .argument }
 
   public init(_ name: String, description: String? = nil, validator: Validator? = nil) {
@@ -65,7 +65,6 @@ public class Argument<T : ArgumentConvertible> : ArgumentDescriptor {
   public let name: String
   public let description: String?
   public let validator: Validator?
-
   public var type: ArgumentType { return .argument }
 
   public init(_ name: String, description: String? = nil, validator: Validator? = nil) {
@@ -99,89 +98,105 @@ public class Option<T : ArgumentConvertible> : ArgumentDescriptor {
   public typealias Validator = (ValueType) throws -> ValueType
 
   public let name: String
+  public let `default`: ValueType
   public let flag: Character?
   public let description: String?
-  public let `default`: ValueType
-  public var type: ArgumentType { return .option }
   public let validator: Validator?
+  public var type: ArgumentType { return .option }
 
-  public init(_ name: String, _ default: ValueType, flag: Character? = nil, description: String? = nil, validator: Validator? = nil) {
+  public init(_ name: String, default: ValueType, flag: Character? = nil, description: String? = nil, validator: Validator? = nil) {
     self.name = name
+    self.`default` = `default`
     self.flag = flag
     self.description = description
-    self.`default` = `default`
     self.validator = validator
+  }
+  
+  @available(*, deprecated, message: "use init(_:default:flag:description:validator:) instead")
+  public convenience init(_ name: String, _ default: ValueType, flag: Character? = nil, description: String? = nil, validator: Validator? = nil) {
+    self.init(name, default: `default`, flag: flag, description: description, validator: validator)
   }
 
   public func parse(_ parser: ArgumentParser) throws -> ValueType {
-    if let value = try parser.shiftValueForOption(name) {
-      let value = try T(string: value)
-
-      if let validator = validator {
-        return try validator(value)
-      }
-
-      return value
+    guard let shifted = try parser.shiftValue(for: name, or: flag) else { return `default` }
+    let value = try T(string: shifted)
+        
+    if let validator = validator {
+      return try validator(value)
     }
-
-    if let flag = flag {
-      if let value = try parser.shiftValueForFlag(flag) {
-        let value = try T(string: value)
-
-        if let validator = validator {
-          return try validator(value)
-        }
-
-        return value
-      }
-    }
-
-    return `default`
+    
+    return value
   }
 }
 
 
 public class Options<T : ArgumentConvertible> : ArgumentDescriptor {
   public typealias ValueType = [T]
+  public typealias Validator = (ValueType) throws -> ValueType
 
   public let name: String
-  public let description: String?
-  public let count: Int
   public let `default`: ValueType
+  public let flag: Character?
+  public let count: Int
+  public let description: String?
+  public let validator: Validator?
   public var type: ArgumentType { return .option }
 
-  public init(_ name: String, _ default: ValueType, count: Int, description: String? = nil) {
+  public init(_ name: String, default: ValueType, flag: Character? = nil, count: Int, description: String? = nil, validator: Validator? = nil) {
     self.name = name
     self.`default` = `default`
+    self.flag = flag
     self.count = count
     self.description = description
+    self.validator = validator
+  }
+  
+  @available(*, deprecated, message: "use init(_:default:flag:count:description:validator:) instead")
+  public convenience init(_ name: String, _ default: ValueType, flag: Character? = nil, count: Int, description: String? = nil, validator: Validator? = nil) {
+    self.init(name, default: `default`, flag: flag, count: count, description: description, validator: validator)
   }
 
   public func parse(_ parser: ArgumentParser) throws -> ValueType {
-    let values = try parser.shiftValuesForOption(name, count: count)
-    return try values?.map { try T(string: $0) } ?? `default`
+    guard let shifted = try parser.shiftValues(for: name, or: flag, count: count) else { return `default` }
+    let values = try shifted.map { try T(string: $0) }
+    
+    if let validator = validator {
+      return try validator(values)
+    }
+    
+    return values
   }
 }
 
 
 public class VariadicOption<T : ArgumentConvertible> : ArgumentDescriptor {
   public typealias ValueType = [T]
+  public typealias Validator = (ValueType) throws -> ValueType
 
   public let name: String
-  public let description: String?
   public let `default`: ValueType
+  public let flag: Character?
+  public let description: String?
+  public let validator: Validator?
   public var type: ArgumentType { return .option }
 
-  public init(_ name: String, _ default: ValueType = [], description: String? = nil) {
+  public init(_ name: String, default: ValueType = [], flag: Character? = nil, description: String? = nil, validator: Validator? = nil) {
     self.name = name
     self.`default` = `default`
+    self.flag = flag
     self.description = description
+    self.validator = validator
+  }
+  
+  @available(*, deprecated, message: "use init(_:default:flag:description:validator:) instead")
+  public convenience init(_ name: String, _ default: ValueType, flag: Character? = nil, description: String? = nil, validator: Validator? = nil) {
+    self.init(name, default: `default`, flag: flag, description: description, validator: validator)
   }
 
   public func parse(_ parser: ArgumentParser) throws -> ValueType {
     var values: ValueType? = nil
 
-    while let shifted = try parser.shiftValueForOption(name) {
+    while let shifted = try parser.shiftValue(for: name, or: flag) {
       let argument = try T(string: shifted)
 
       if values == nil {
@@ -190,6 +205,10 @@ public class VariadicOption<T : ArgumentConvertible> : ArgumentDescriptor {
       values?.append(argument)
     }
 
+    if let validator = validator, let values = values {
+      return try validator(values)
+    }
+    
     return values ?? `default`
   }
 }
@@ -199,20 +218,25 @@ public class Flag : ArgumentDescriptor {
   public typealias ValueType = Bool
 
   public let name: String
+  public let `default`: ValueType
   public let flag: Character?
   public let disabledName: String
   public let disabledFlag: Character?
   public let description: String?
-  public let `default`: ValueType
   public var type: ArgumentType { return .option }
 
-  public init(_ name: String, flag: Character? = nil, disabledName: String? = nil, disabledFlag: Character? = nil, description: String? = nil, default: Bool = false) {
+  public init(_ name: String, default: Bool = false, flag: Character? = nil, disabledName: String? = nil, disabledFlag: Character? = nil, description: String? = nil) {
     self.name = name
+    self.`default` = `default`
     self.disabledName = disabledName ?? "no-\(name)"
     self.flag = flag
     self.disabledFlag = disabledFlag
     self.description = description
-    self.`default` = `default`
+  }
+  
+  @available(*, deprecated, message: "use init(_:default:flag:disabledName:disabledFlag:description:) instead")
+  public convenience init(_ name: String, _ default: ValueType, flag: Character? = nil, disabledName: String? = nil, disabledFlag: Character? = nil, description: String? = nil) {
+    self.init(name, default: `default`, flag: flag, disabledName: disabledName, disabledFlag: disabledFlag, description: description)
   }
 
   public func parse(_ parser: ArgumentParser) throws -> ValueType {
@@ -245,6 +269,31 @@ class BoxedArgumentDescriptor {
   let description: String?
   let `default`: String?
   let type: ArgumentType
+  let flag: Character?
+
+  var summary: String {
+    var line = ""
+
+    switch type {
+    case .argument:
+      line += "\(name)"
+    case .option:
+      if let flag = flag {
+        line += "-\(flag), "
+      }
+      line += "--\(name)"
+    }
+
+    if let `default` = `default` {
+      line += " [default: \(`default`)]"
+    }
+
+    if let description = description {
+      line += " - \(description)"
+    }
+
+    return line
+  }
 
   init<T : ArgumentDescriptor>(value: T) {
     name = value.name
@@ -253,13 +302,23 @@ class BoxedArgumentDescriptor {
 
     if let value = value as? Flag {
       `default` = value.`default`.description
+      flag = value.flag
     } else if let value = value as? Option<String> {
       `default` = value.`default`.description
+      flag = nil
     } else if let value = value as? Option<Int> {
       `default` = value.`default`.description
+      flag = nil
     } else {
-      // TODO, default for Option of generic type
-      `default` = nil
+      let mirror = Mirror(reflecting: value)
+
+      if let defaultDescendant = mirror.descendant("default"), let defaultConvertible = defaultDescendant as? CustomStringConvertible {
+        `default` = defaultConvertible.description
+      } else {
+        `default` = nil
+      }
+
+      flag = nil
     }
   }
 }
@@ -321,7 +380,8 @@ class Help : Error, ANSIConvertible, CustomStringConvertible {
     if let group = group {
       output.append("Commands:")
       output.append("")
-      for command in group.commands {
+      let commands = group.commands.sorted(by: { $0.name < $1.name })
+      for command in commands {
         if let description = command.description {
           output.append("    + \(command.name) - \(description)")
         } else {
@@ -332,32 +392,14 @@ class Help : Error, ANSIConvertible, CustomStringConvertible {
     } else if !arguments.isEmpty {
       output.append("Arguments:")
       output.append("")
-
-      output += arguments.map { argument in
-        if let description = argument.description {
-          return "    \(argument.name) - \(description)"
-        } else {
-          return "    \(argument.name)"
-        }
-      }
-
+      output += arguments.map { "    \($0.summary)" }
       output.append("")
     }
 
     if !options.isEmpty {
       output.append("Options:")
       for option in options {
-        var line = "    --\(option.name)"
-
-        if let `default` = option.default {
-          line += " [default: \(`default`)]"
-        }
-
-        if let description = option.description {
-          line += " - \(description)"
-        }
-
-        output.append(line)
+        output.append("    \(option.summary)")
       }
     }
 
@@ -383,7 +425,8 @@ class Help : Error, ANSIConvertible, CustomStringConvertible {
     if let group = group {
       output.append("Commands:")
       output.append("")
-      for command in group.commands {
+      let commands = group.commands.sorted(by: { $0.name < $1.name })
+      for command in commands {
         if let description = command.description {
           output.append("    + \(ANSI.green)\(command.name)\(ANSI.reset) - \(description)")
         } else {
@@ -409,7 +452,13 @@ class Help : Error, ANSIConvertible, CustomStringConvertible {
     if !options.isEmpty {
       output.append("Options:")
       for option in options {
-        var line = "    \(ANSI.blue)--\(option.name)\(ANSI.reset)"
+        var line = "    "
+
+        if let flag = option.flag {
+          line += "\(ANSI.blue)-\(flag)\(ANSI.reset), "
+        }
+
+        line += "\(ANSI.blue)--\(option.name)\(ANSI.reset)"
 
         if let `default` = option.default {
           line += " [default: \(`default`)]"