+++ Pods/Commander/Sources/Commander/ArgumentParser.swift
@@ -1,4 +1,4 @@
-private enum Arg : CustomStringConvertible {
+private enum Arg : CustomStringConvertible, Equatable {
   /// A positional argument
   case argument(String)
 
@@ -6,7 +6,7 @@ private enum Arg : CustomStringConvertible {
   case option(String)
 
   /// A flag
-  case flag(Set<Character>)
+  case flag([Character])
 
   var description:String {
     switch self {
@@ -32,6 +32,11 @@ private enum Arg : CustomStringConvertible {
 }
 
 
+private func == (lhs: Arg, rhs: Arg) -> Bool {
+  return lhs.description == rhs.description
+}
+
+
 public struct ArgumentParserError : Error, Equatable, CustomStringConvertible {
   public let description: String
 
@@ -45,26 +50,40 @@ public func ==(lhs: ArgumentParserError, rhs: ArgumentParserError) -> Bool {
   return lhs.description == rhs.description
 }
 
-
 public final class ArgumentParser : ArgumentConvertible, CustomStringConvertible {
   fileprivate var arguments:[Arg]
 
+  public typealias Option = String
+  public typealias Flag = Character
+
   /// Initialises the ArgumentParser with an array of arguments
   public init(arguments: [String]) {
-    self.arguments = arguments.map { argument in
-      if argument.hasPrefix("-") {
+    let splitArguments = arguments.split(maxSplits: 1, omittingEmptySubsequences: false) { $0 == "--" }
+
+    let unfixedArguments: [String]
+    let fixedArguments: [String]
+    if splitArguments.count == 2, let prefix = splitArguments.first, let suffix = splitArguments.last {
+      unfixedArguments = Array(prefix)
+      fixedArguments = Array(suffix)
+    } else {
+      unfixedArguments = arguments
+      fixedArguments = []
+    }
+
+    self.arguments = unfixedArguments.map { argument in
+      if argument.first == "-" {
         let flags = argument[argument.index(after: argument.startIndex)..<argument.endIndex]
 
-        if flags.hasPrefix("-") {
+        if flags.first == "-" {
           let option = flags[flags.index(after: flags.startIndex)..<flags.endIndex]
           return .option(String(option))
         }
-
-        return .flag(Set(flags))
+        return .flag(Array(String(flags)))
       }
 
       return .argument(argument)
     }
+    self.arguments.append(contentsOf: fixedArguments.map { .argument($0) })
   }
 
   public init(parser: ArgumentParser) throws {
@@ -75,8 +94,8 @@ public final class ArgumentParser : ArgumentConvertible, CustomStringConvertible
     return arguments.map { $0.description }.joined(separator: " ")
   }
 
-  public var isEmpty:Bool {
-    return arguments.isEmpty
+  public var isEmpty: Bool {
+    return arguments.first { $0 != .argument("") } == nil
   }
 
   public var remainder:[String] {
@@ -100,12 +119,12 @@ public final class ArgumentParser : ArgumentConvertible, CustomStringConvertible
   }
 
   /// Returns the value for an option (--name Kyle, --name=Kyle)
-  public func shiftValueForOption(_ name:String) throws -> String? {
-    return try shiftValuesForOption(name)?.first
+  public func shiftValue(for name: Option) throws -> String? {
+    return try shiftValues(for: name)?.first
   }
 
   /// Returns the value for an option (--name Kyle, --name=Kyle)
-  public func shiftValuesForOption(_ name:String, count:Int = 1) throws -> [String]? {
+  public func shiftValues(for name: Option, count: Int = 1) throws -> [String]? {
     var index = 0
     var hasOption = false
 
@@ -147,7 +166,7 @@ public final class ArgumentParser : ArgumentConvertible, CustomStringConvertible
   }
 
   /// Returns whether an option was specified in the arguments
-  public func hasOption(_ name:String) -> Bool {
+  public func hasOption(_ name: Option) -> Bool {
     var index = 0
     for argument in arguments {
       switch argument {
@@ -167,14 +186,14 @@ public final class ArgumentParser : ArgumentConvertible, CustomStringConvertible
   }
 
   /// Returns whether a flag was specified in the arguments
-  public func hasFlag(_ flag:Character) -> Bool {
+  public func hasFlag(_ flag: Flag) -> Bool {
     var index = 0
     for argument in arguments {
       switch argument {
       case .flag(let option):
         var options = option
         if options.contains(flag) {
-          options.remove(flag)
+          options.removeAll(where: { $0 == flag })
           arguments.remove(at: index)
 
           if !options.isEmpty {
@@ -193,12 +212,12 @@ public final class ArgumentParser : ArgumentConvertible, CustomStringConvertible
   }
 
   /// Returns the value for a flag (-n Kyle)
-  public func shiftValueForFlag(_ flag:Character) throws -> String? {
-    return try shiftValuesForFlag(flag)?.first
+  public func shiftValue(for flag: Flag) throws -> String? {
+    return try shiftValues(for: flag)?.first
   }
 
   /// Returns the value for a flag (-n Kyle)
-  public func shiftValuesForFlag(_ flag:Character, count:Int = 1) throws -> [String]? {
+  public func shiftValues(for flag: Flag, count: Int = 1) throws -> [String]? {
     var index = 0
     var hasFlag = false
 
@@ -237,6 +256,28 @@ public final class ArgumentParser : ArgumentConvertible, CustomStringConvertible
       }
     }
 
+    return nil
+  }
+  
+  /// Returns the value for an option (--name Kyle, --name=Kyle) or flag (-n Kyle)
+  public func shiftValue(for name: Option, or flag: Flag?) throws -> String? {
+    if let value = try shiftValue(for: name) {
+      return value
+    } else if let flag = flag, let value = try shiftValue(for: flag) {
+      return value
+    }
+    
+    return nil
+  }
+  
+  /// Returns the values for an option (--name Kyle, --name=Kyle) or flag (-n Kyle)
+  public func shiftValues(for name: Option, or flag: Flag?, count: Int = 1) throws -> [String]? {
+    if let value = try shiftValues(for: name, count: count) {
+      return value
+    } else if let flag = flag, let value = try shiftValues(for: flag, count: count) {
+      return value
+    }
+    
     return nil
   }
 }